requirements.txt
Flask==1.1.1
Flask-WTF==0.14.3
Flask-Mail==0.9.1
email-validator==1.1.1
folium==0.11.0
branca==0.4.1
stravaio==0.0.9
gunicorn==20.0.4
opencensus-ext-flask==0.7.3
opencensus-ext-azure==1.0.3
